osposobljena komunikacija sa back-om

+++ UserFront/src/components/ShoppingCart.vue
@@ -7,11 +7,11 @@
     </div>
         
     <div class='container' id='search'>
-      <p class="cart-empty" v-if="oglasi_korpa.length == 0">Your Shopping Cart is Empty!</p>
+      <h3 class="cart-empty" v-if="oglasi_korpa.length == 0">Vaša korpa je prazna!</h3>
       <div class="row" >
         <div  class="col-lg-3 col-md-4 col-sm-6 mb-4" v-bind:key='oglas.id'  v-for="oglas in oglasi_korpa">
           <div class="card h-100 ">
-              <img id='overview-img' class="card-img-top" v-bind:src="oglas.automobil.images[0]" alt="">
+              <img id='overview-img' class="card-img-top" v-bind:src="oglas.automobil.slikeVozila.slika[0]" alt="">
               <div class="card-body">
                   <div class="card-text">
                       <h5 class="theme-colr"><b>Automobil: </b>{{oglas.automobil.markaAut}} {{oglas.automobil.modelAut}}</h5> 
@@ -48,7 +48,9 @@
 </template>
 
 <script>
-  export default {
+import UserDataService from '../services/UserDataService'
+import StarRating from "vue-star-rating";
+export default {
     data() {
       return {
         total:0,
@@ -62,180 +64,13 @@
               odDatuma:null,
               doDatuma:null,
             },
-            {
-              agentId:null,
-              userId:null,
-              oglasId:null,
-              odDatuma:null,
-              doDatuma:null,
-            },
+            ...
           ]
         */
        
-        oglasi_korpa:[  //nalaze se samo oglasi zarad prikaza u korpi i manipulaciju njima
-          {
-            // oglas
-              id:1,
-              odDatuma:'25.5.2020',
-              doDatuma:'25.6.2020',
-              TAdresa:{
-                mesto:'Novi Sad',
-                ulica:'9. Marta',
-                broj:'bb',
-                postanskiBroj:'21000',
-                longitude:'45',
-                latitude:'54',
-              },
-              planiranaKilometraza:2000,
-              agentId:2,  
-              korisnickoIme:'Test', //u DTOu za korisnika koji je kreirao oglas.
-              //automobil
-              automobil:{
-                id:'1',
-                markaAut:'BMW 1',
-                modelAut:'M5',
-                klasaAut:'SUV',
-                vrstaGoriva:'dizel',
-                tipMenjaca:'manuelni',
-                ukupnaOcena:5,
-                brojSedistaZaDecu:1,
-                predjenaKilometraza:5000,
-                collisionDamageWaiver:true,
-                images:['https://source.unsplash.com/RCAhiGJsUUE/1920x1080','https://source.unsplash.com/wfh8dDlNFOk/1920x1080','https://source.unsplash.com/O7fzqFEfLlo/1920x1080'],
-              },
-              //cena    
-              cenovnik:{
-                id:'1',
-                cenaPoDanu:100,
-                nazivCenovnika:'Cenovnik 1',
-                popustZaPreko30Dana:10,
-                cenaCollisionDamageWaiver:1000,
-                cenaPoKilometru:10  
-              },
-            },
-          {
-            //oglas
-            id:3,
-            odDatuma:'18.6.2020',
-            doDatuma:'25.7.2020',
-            TAdresa:{
-              mesto:'Beograd',
-              ulica:'Ne znanog i znanog junaka',
-              broj:'bb',
-              postanskiBroj:'11000',
-              longitude:'21.23',
-              latitude:'34.14',
-            },
-            planiranaKilometraza:2500,
-            agentId:1, 
-            korisnickoIme:'Other host', //u DTOu za korisnika koji je kreirao oglas.
-            //Oglas/Automobil
-            automobil:{
-              id:3,
-              markaAut:'Audi 2',
-              modelAut:'A6',
-              klasaAut:'Gradski',
-              vrstaGoriva:'dizel',
-              tipMenjaca:'manuelni',
-              predjenaKilometraza:4500,
-              ukupnaOcena:4.25,
-              collisionDamageWaiver:true,
-              brojSedistaZaDecu:1,
-              images:['https://source.unsplash.com/O7fzqFEfLlo/1920x1080'],
-            },
-            //Oglas/Cenovnik
-            cenovnik:{ 
-              id:'2',
-              cenaPoDanu:500,
-              nazivCenovnika:'Cenovnik 2',
-              popustZaPreko30Dana:20,
-              cenaCollisionDamageWaiver:null,
-              cenaPoKilometru:20
-            },
-          },
-          {
-            //oglas
-            id:4,
-            odDatuma:'25.5.2020',
-            doDatuma:'15.6.2020',
-            TAdresa:{
-              mesto:'Novi Sad',
-              ulica:'19. Juna',
-              broj:'45',
-              postanskiBroj:'11000',
-              longitude:'12',
-              latitude:'21',
-            },
-            planiranaKilometraza:3000,
-            agentId:2, 
-            korisnickoIme:'Test', //u DTOu za korisnika koji je kreirao oglas.
-            //automobil
-            automobil:{
-              id:'3',
-              markaAut:'Audi 3',
-              modelAut:'A6',
-              klasaAut:'Gradski auto',
-              vrstaGoriva:'dizel',
-              tipMenjaca:'manuelni',
-              ukupnaOcena:3.12,
-              brojSedistaZaDecu:2,
-              predjenaKilometraza:650,
-              collisionDamageWaiver:false,
-              images:['https://source.unsplash.com/wfh8dDlNFOk/1920x1080','https://source.unsplash.com/O7fzqFEfLlo/1920x1080'],
-            },
-            cenovnik:{
-              id:'3',
-              cenaPoDanu:300,
-              nazivCenovnika:'Cenovnik 3',
-              popustZaPreko30Dana:null,
-              cenaCollisionDamageWaiver:3000,
-              cenaPoKilometru:30
-            },
-          },
-           {
-            //oglas
-            id:6,
-            odDatuma:'18.6.2020',
-            doDatuma:'25.7.2020',
-            TAdresa:{
-              mesto:'Beograd',
-              ulica:'Ne znanog i znanog junaka',
-              broj:'bb',
-              postanskiBroj:'11000',
-              longitude:'21.23',
-              latitude:'34.14',
-            },
-            planiranaKilometraza:2500,
-            agentId:2, 
-            korisnickoIme:'Other host', //u DTOu za korisnika koji je kreirao oglas.
-            //Oglas/Automobil
-            automobil:{
-              id:1,
-              markaAut:'Audi 4',
-              modelAut:'A8',
-              klasaAut:'Gradski',
-              vrstaGoriva:'dizel',
-              tipMenjaca:'manuelni',
-              predjenaKilometraza:5500,
-              ukupnaOcena:3.25,
-              collisionDamageWaiver:true,
-              brojSedistaZaDecu:1,
-              images:['https://source.unsplash.com/O7fzqFEfLlo/1920x1080'],
-            },
-            //Oglas/Cenovnik
-            cenovnik:{ 
-              id:'2',
-              cenaPoDanu:585,
-              nazivCenovnika:'Cenovnik 2',
-              popustZaPreko30Dana:50,
-              cenaCollisionDamageWaiver:null,
-              cenaPoKilometru:20
-            },
-          },
-        ],
+        oglasi_korpa:[],  //nalaze se samo oglasi zarad prikaza u korpi i manipulaciju njima
         rezervacije:[],
         poruke:[],
- 
       }
     },
     methods:{
@@ -249,14 +84,47 @@
         }
         return false;
       },
+    //REMOVE Logika
       removeAds:function(id){
-        alert(`Oglas ${id} ce biti uklonjen!`);
-        // Parsirati string iz localStorage u JSON (vec odradjeno this.korpa)
-        // Obrisati element koji zelimo (with slice() )
-        this.korpa = this.removeItem(this.korpa, id);
-        // Konvertovati nazad JSON u string i ponovo ga vratiti u localStorage
-        localStorage.setItem('cart', JSON.stringify(this.korpa));
+        if (confirm('Da li ste sigurni da želite ukloniti ovu rezervaciju?')) { 
+          this.korpa = this.removeItem(this.korpa, id);
+  
+          localStorage.setItem('cart', JSON.stringify(this.korpa));
+        }
       },
+      //Pomocna metoda koja se poziva iz metode removeAds
+      //Sluzi za proveru da li je neki oglas u nekom nizu
+      //Ako jeste brise se iz niza i vraca se niz...
+      removeItem(korpa, oglasId) {
+        console.log('Usao u removeItem');
+        for(let i = 0; i< korpa.length; i ++){
+          if(korpa[i].oglasId == oglasId){
+            console.log('Obrisano: ' + korpa.splice(i, 1));
+            this.removeVisually(oglasId);
+          }
+        }
+        return korpa;
+      },
+      //Pomocna metoda koja se poziva iz metode  removeItem
+      //Sluzi da u slucaju kada se brise iz korpe neki element
+      //se on obirise i vizuelno uklanjanjem tog oglasa iz oglas_korpa
+      removeVisually(oglasId){
+        console.log('Usao u removeVisually!');
+          for(let i = 0; i< this.oglasi_korpa.length; i ++){
+            if(this.oglasi_korpa[i].id == oglasId){
+              console.log('sadadasdsadasdsa');
+              this.oglasi_korpa.splice(i, 1);
+              this.total = 0;
+              for(let i = 0; i < this.oglasi_korpa.length; i++){
+
+                this.countTotalPrice(this.oglasi_korpa[i]);
+              }
+            }
+          }
+      },
+    //<--remove
+
+    //ADD LOGIKA
       //Metoda koja sluzi za kreiranje rezervacija i njihovo slanje na bek
       makeReseravation:function(id){
         let bundle = false;
@@ -295,16 +163,7 @@
           }
         }
       },
-      //Metoda koja sluzi za prekid rezervacije i brisanje sadrzaja korpe iz localStorage 
-      cancelReservation:function(){
-        if (confirm('Da li ste sigurni da želite odustati? Vaša korpa će biti prazna!')) {
-          localStorage.removeItem('cart');
-          this.$router.push(`/cars`);
-        }
-      },
-      goBack:function(){
-        this.$router.push(`/cars`);
-      },
+
       //Metoda koja sluzi za kreiranje rezervacije sa samo jednom narudzbenicom...
       createSingleReservation:function(oglas){
         for(let i = 0; i < this.korpa.length; i++){
@@ -367,20 +226,17 @@
             }
           }
         }
-          const rezervacijaOglasDTO = {
-          //rezervacija
-            ukupnaCena: ukupnaCena,
-            statusRezervacije:'PENDING',
-            username:'Happy User 2', //u rezervDTOu za korisnika koji je rezervisao oglas.
-            bundle:true,
-            poruka:'',
-            narudzbenice:narudzbenice, //u bundleReserv je array narudzbenica...
-          //Na beku se dodaje:
-            //commonDataId: u oba
-            //NarudzbenicaId: u rezervaciju
-          }
-          this.rezervacije.push(rezervacijaOglasDTO);
-          localStorage.setItem('reserv', JSON.stringify(this.rezervacije));
+        const rezervacijaOglasDTO = {
+        //rezervacija
+          ukupnaCena: ukupnaCena,
+          statusRezervacije:'PENDING',
+          username:'Happy User 2', //u rezervDTOu za korisnika koji je rezervisao oglas.
+          bundle:true,
+          poruka:'',
+          narudzbenice:narudzbenice, //u bundleReserv je array narudzbenica...
+        }
+        this.rezervacije.push(rezervacijaOglasDTO);
+        localStorage.setItem('reserv', JSON.stringify(this.rezervacije));
       },
     
       //Metoda koja sluzi da razvrsta oglase na one koji su u nekom bundle i one koji su sami.
@@ -431,34 +287,7 @@
         console.log(`singleAds: ${singleAds.length}`);
         return singleAds;
       },
-      //Pomocna metoda koja se poziva iz metode removeAds
-      //Sluzi za proveru da li je neki oglas u nekom nizu
-      //Ako jeste brise se iz niza i vraca se niz...
-      removeItem(korpa, oglasId) {
-        console.log('Usao u removeItem');
-        for(let i = 0; i< korpa.length; i ++){
-          if(korpa[i].oglasId == oglasId){
-            console.log('Obrisano: ' + korpa.splice(i, 1));
-            this.removeVisually(oglasId);
-          }
-        }
-        return korpa;
-      },
-      //Pomocna metoda koja se poziva iz metode  removeItem
-      //Sluzi da u slucaju kada se brise iz korpe neki element
-      //se on obirise i vizuelno uklanjanjem tog oglasa iz oglas_korpa
-      removeVisually(oglasId){
-        console.log('Usao u removeVisually!');
-          for(let i = 0; i< this.oglasi_korpa.length; i ++){
-            if(this.oglasi_korpa[i].id == oglasId){
-              console.log('sadadasdsadasdsa');
-              this.oglasi_korpa.splice(i, 1);
-            }
-          }
-      },
-      showDetails:function(id){
-        this.$router.push(`/reservation/${id}/details`);
-      },
+
       //Pomocna metoda koja se poziva iz metode putInBundle
       //Sluzi za proveru da li je neki oglas vec u bundlu
       //Ako jeste vraca se true, ako nije false
@@ -470,6 +299,26 @@
         }
         return false;
       },
+    //<--add
+
+    //BRISANJE, POVRATAK, DETALJI
+      //Metoda koja sluzi za prekid rezervacije i brisanje sadrzaja korpe iz localStorage 
+      cancelReservation:function(){
+        if (confirm('Da li ste sigurni da želite odustati? Vaša korpa će biti prazna!')) {
+          localStorage.removeItem('cart');
+          this.$router.push(`/cars`);
+        }
+      },
+      goBack:function(){
+        this.$router.push(`/cars`);
+      },
+
+      showDetails:function(id){
+        this.$router.push(`/shoppingCart/reservation/${id}/details`);
+      },
+    //<--Del,back,detail
+
+    //POMOCNE METODE ZA DINAMICKU KALKULACIJU 
       //Pomocna metoda za racunanje broja dana na osnovu pocetnog i krajnjeg datuma u narudzbenici...
       calculateNoOfDays(odDatuma,doDatuma){
         //Konvertovanje datuma u broj dana... 
@@ -504,38 +353,60 @@
         }
         return ukupnaCena;
       },
-      countTotalPrice:function(){
-        for(let i = 0; i< this.oglasi_korpa.length; i++ ){
+      countTotalPrice:function(oglasi_korpa){
+        console.log("LENGTH: " + JSON.stringify(oglasi_korpa));
+        // for(let i = 0; i< oglasi_korpa.length; i++ ){
           for(let j = 0; j < this.korpa.length; j++ ){
-            if(this.oglasi_korpa[i].id == this.korpa[j].oglasId){
-              console.log(`Nasli se u ${this.oglasi_korpa[i].id} i ${this.korpa[j].oglasId}`);
+            if(oglasi_korpa.id == this.korpa[j].oglasId){
+              console.log(`Nasli se u ${oglasi_korpa.id} i ${this.korpa[j].oglasId}`);
               const odabran_br_dana = this.calculateNoOfDays(this.korpa[j].odDatuma , this.korpa[j].doDatuma);
-              this.total += this.calculatePrice(this.oglasi_korpa[j].cenovnik.cenaPoDanu, this.oglasi_korpa[j].cenovnik.popustZaPreko30Dana, odabran_br_dana);
+              this.total += this.calculatePrice(oglasi_korpa.cenovnik.cenaPoDanu, oglasi_korpa.cenovnik.popustZaPreko30Dana, odabran_br_dana);
               continue;
             }
           }
-        }
+        // }
       },
+      /* countNewTotalPrice:function() {
+          //Preuzimanje objekta korpa iz localStorage
+          this.korpa = JSON.parse(localStorage.getItem('cart'));
+          for(let i = 0; i < this.korpa.length; i++){
+          //Traze se svi oglasi iz korpe preko id-ja kako bi se smestili u odabraniOglasi 
+          //i prikazali njihovi podacu i slike vizuelno u korpi...
+          UserDataService.getOglasDetails(this.korpa[i].oglasId).then(response => {
+            this.oglasi_korpa = [];
+            this.oglasi_korpa.push(response.data);  
+            this.countTotalPrice(response.data);
+          });
+        }
+      }*/ 
+    //<--calc
+    },
+      components: {
+        "star-rating": StarRating,
     },
     created(){
       if(JSON.parse(localStorage.getItem('token')) == null){
-           this.$router.push(`/login`);
+          this.$router.push(`/login`);
+      }
+      else{
+
+        if(JSON.parse(localStorage.getItem('cart'))== null){
+          this.korpa = [];
         }
-        else{
+        else {
           // Preuzimanje objekta korpa iz localStorage
           this.korpa = JSON.parse(localStorage.getItem('cart'));
-          
-          // this.poruke = JSON.parse(localStorage.getItem('messages'));
-          let oglasiIzKorpe = [];
           for(let i = 0; i < this.korpa.length; i++){
-            //ili preko urla ih poslati...
-            oglasiIzKorpe.push(this.korpa[i].oglasId);
             //Traze se svi oglasi iz korpe preko id-ja kako bi se smestili u odabraniOglasi 
             //i prikazali njihovi podacu i slike vizuelno u korpi...
-            //getAdsById(oglasiIzKorpe).response(this.oglasi_korpa = response)....  
-      }
-      this.countTotalPrice();
+            UserDataService.getOglasDetails(this.korpa[i].oglasId).then(response => {
+              this.oglasi_korpa.push(response.data);  
+              this.countTotalPrice(response.data);
+            });
+          }
         }
+        // this.countTotalPrice(this.oglasi_korpa);
+      }
       
     },
   }
